refactor: ManagerMain 페이지에서 MapDrawer 컴포넌트 분리
